+++ web/src/app/dashboard/components/BroadcastPanel.tsx
@@ -1,18 +1,12 @@
 import {
-  ArrowDownIcon,
-  ArrowUpIcon,
-  Bars3BottomLeftIcon,
   CheckCircleIcon,
   ClockIcon,
   ExclamationCircleIcon,
   PaperAirplaneIcon,
   PencilSquareIcon,
-  PhotoIcon,
-  PlusIcon,
   TrashIcon,
-  XMarkIcon,
 } from "@heroicons/react/24/outline";
-import { useCallback, useRef, useState } from "react";
+import { useState } from "react";
 
 import {
   useBroadcasts,
@@ -22,9 +16,10 @@ import {
 import { useInfiniteScroll } from "~/hooks/useInfiniteScroll";
 import { confirmDialog } from "~/lib/dialog";
 import { formatDateTime } from "~/lib/format";
-import type { BroadcastMessage, BroadcastPart, BroadcastStatus } from "~/types";
-import { getImageUrl, type PartState } from "./broadcast-helpers";
-import { MAX_PARTS, type ModalMode, useBroadcastForm } from "./useBroadcastForm";
+import type { BroadcastMessage, BroadcastStatus } from "~/types";
+import { BroadcastFormModal } from "./broadcast/BroadcastFormModal";
+import { BroadcastPartPreview } from "./broadcast/BroadcastPartPreview";
+import type { ModalMode } from "./broadcast/useBroadcastForm";
 
 const STATUS_LABELS: Record<BroadcastStatus, string> = {
   draft: "下書き",
@@ -45,387 +40,6 @@ const STATUS_ACTIVE_STYLES: Record<BroadcastStatus, string> = {
   scheduled: "bg-blue-600 text-white",
   sent: "bg-green-600 text-white",
   failed: "bg-red-600 text-white",
-};
-
-type BroadcastFormModalProps = {
-  mode: ModalMode;
-  broadcast?: BroadcastMessage;
-  onClose: () => void;
-};
-
-const PartEditor = ({
-  part,
-  index,
-  total,
-  onChange,
-  onRemove,
-  onMove,
-}: {
-  part: PartState;
-  index: number;
-  total: number;
-  onChange: (index: number, part: PartState) => void;
-  onRemove: (index: number) => void;
-  onMove: (index: number, direction: "up" | "down") => void;
-}) => {
-  const fileInputRef = useRef<HTMLInputElement>(null);
-
-  const hasContent =
-    part.type === "text"
-      ? part.text.trim().length > 0
-      : !!(part.imageR2Key || part.file);
-  const canSwitchToText = part.type === "text" || !hasContent;
-  const canSwitchToImage = part.type === "image" || !hasContent;
-
-  const handleFileSelect = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const file = e.target.files?.[0];
-    if (!file) return;
-    const previewUrl = URL.createObjectURL(file);
-    onChange(index, {
-      id: part.id,
-      type: "image",
-      imageR2Key: "",
-      file,
-      previewUrl,
-    });
-  };
-
-  return (
-    <div className="border border-stone-200 rounded-lg overflow-hidden">
-      <div className="flex items-center justify-between bg-stone-50 px-2 py-1.5">
-        <div className="flex">
-          <button
-            type="button"
-            disabled={!canSwitchToText}
-            onClick={() =>
-              part.type !== "text" &&
-              onChange(index, { id: part.id, type: "text", text: "" })
-            }
-            className={`flex items-center gap-1 px-2.5 py-1 text-xs font-medium rounded-l border transition-colors ${
-              part.type === "text"
-                ? "bg-white text-stone-800 border-stone-300"
-                : canSwitchToText
-                  ? "bg-stone-100 text-stone-400 border-stone-200 hover:text-stone-600"
-                  : "bg-stone-100 text-stone-300 border-stone-200 cursor-not-allowed"
-            }`}
-          >
-            <Bars3BottomLeftIcon className="w-3.5 h-3.5" />
-            テキスト
-          </button>
-          <button
-            type="button"
-            disabled={!canSwitchToImage}
-            onClick={() =>
-              part.type !== "image" &&
-              onChange(index, {
-                id: part.id,
-                type: "image",
-                imageR2Key: "",
-              })
-            }
-            className={`flex items-center gap-1 px-2.5 py-1 text-xs font-medium rounded-r border border-l-0 transition-colors ${
-              part.type === "image"
-                ? "bg-white text-stone-800 border-stone-300"
-                : canSwitchToImage
-                  ? "bg-stone-100 text-stone-400 border-stone-200 hover:text-stone-600"
-                  : "bg-stone-100 text-stone-300 border-stone-200 cursor-not-allowed"
-            }`}
-          >
-            <PhotoIcon className="w-3.5 h-3.5" />
-            画像
-          </button>
-        </div>
-        <div className="flex items-center gap-0.5">
-          <button
-            type="button"
-            onClick={() => onMove(index, "up")}
-            disabled={index === 0}
-            className="p-1 text-stone-400 hover:text-stone-600 disabled:opacity-30"
-          >
-            <ArrowUpIcon className="w-3.5 h-3.5" />
-          </button>
-          <button
-            type="button"
-            onClick={() => onMove(index, "down")}
-            disabled={index === total - 1}
-            className="p-1 text-stone-400 hover:text-stone-600 disabled:opacity-30"
-          >
-            <ArrowDownIcon className="w-3.5 h-3.5" />
-          </button>
-          <button
-            type="button"
-            onClick={() => onRemove(index)}
-            disabled={total <= 1}
-            className="p-1 text-red-400 hover:text-red-600 disabled:opacity-30"
-          >
-            <XMarkIcon className="w-3.5 h-3.5" />
-          </button>
-        </div>
-      </div>
-
-      <div className="p-3">
-        {part.type === "text" ? (
-          <div>
-            <textarea
-              value={part.text}
-              onChange={(e) =>
-                onChange(index, {
-                  id: part.id,
-                  type: "text",
-                  text: e.target.value,
-                })
-              }
-              placeholder="テキストを入力"
-              maxLength={5000}
-              rows={4}
-              className="w-full px-3 py-2 border border-stone-300 rounded-lg text-sm focus:outline-none focus:ring-2 focus:ring-teal-500 focus:border-transparent resize-y"
-            />
-            <div className="text-xs text-stone-400 mt-1 text-right">
-              {part.text.length} / 5000
-            </div>
-          </div>
-        ) : (
-          <div>
-            {part.previewUrl || part.imageR2Key ? (
-              <div className="relative">
-                <img
-                  src={part.previewUrl || getImageUrl(part.imageR2Key)}
-                  alt="プレビュー"
-                  className="max-h-48 rounded-lg object-contain mx-auto"
-                />
-                <button
-                  type="button"
-                  onClick={() =>
-                    onChange(index, {
-                      id: part.id,
-                      type: "image",
-                      imageR2Key: "",
-                    })
-                  }
-                  className="absolute top-1 right-1 p-1 bg-black/50 text-white rounded-full hover:bg-black/70"
-                >
-                  <XMarkIcon className="w-4 h-4" />
-                </button>
-              </div>
-            ) : (
-              <button
-                type="button"
-                onClick={() => fileInputRef.current?.click()}
-                className="w-full py-8 border-2 border-dashed border-stone-300 rounded-lg text-stone-400 hover:border-teal-400 hover:text-teal-500 transition-colors flex flex-col items-center gap-2"
-              >
-                <PhotoIcon className="w-8 h-8" />
-                <span className="text-sm font-medium">写真をアップロード</span>
-                <span className="text-xs">JPG, PNG / 10MB以下</span>
-              </button>
-            )}
-            <input
-              ref={fileInputRef}
-              type="file"
-              accept="image/jpeg,image/png"
-              onChange={handleFileSelect}
-              className="hidden"
-            />
-          </div>
-        )}
-      </div>
-    </div>
-  );
-};
-
-const BroadcastFormModal = ({
-  mode,
-  broadcast,
-  onClose,
-}: BroadcastFormModalProps) => {
-  const {
-    parts,
-    timing,
-    setTiming,
-    scheduledAt,
-    setScheduledAt,
-    isPending,
-    isValid,
-    isError,
-    errorMessage,
-    handlePartChange,
-    handlePartRemove,
-    handlePartMove,
-    handleAddPart,
-    handleSubmit,
-  } = useBroadcastForm({ mode, broadcast, onClose });
-
-  return (
-    <div className="fixed inset-0 z-50 flex items-center justify-center">
-      <button
-        type="button"
-        className="absolute inset-0 bg-black/50 cursor-default"
-        onClick={onClose}
-        aria-label="閉じる"
-      />
-      <div className="relative bg-white rounded-xl shadow-xl w-full max-w-2xl mx-4 p-6 max-h-[90dvh] overflow-auto">
-        <div className="flex items-center justify-between mb-4">
-          <h2 className="text-lg font-bold text-stone-800">
-            {mode === "create" ? "新規配信作成" : "配信を編集"}
-          </h2>
-          <button
-            type="button"
-            onClick={onClose}
-            className="p-1 text-stone-400 hover:text-stone-600 rounded-md"
-            aria-label="閉じる"
-          >
-            <XMarkIcon className="w-5 h-5" aria-hidden="true" />
-          </button>
-        </div>
-
-        <div className="space-y-3">
-          {parts.map((part, index) => (
-            <PartEditor
-              key={part.id}
-              part={part}
-              index={index}
-              total={parts.length}
-              onChange={handlePartChange}
-              onRemove={handlePartRemove}
-              onMove={handlePartMove}
-            />
-          ))}
-
-          {parts.length < MAX_PARTS && (
-            <button
-              type="button"
-              onClick={handleAddPart}
-              className="w-full py-2 border border-dashed border-teal-300 rounded-lg text-teal-600 hover:bg-teal-50 transition-colors text-sm font-medium flex items-center justify-center gap-1"
-            >
-              <PlusIcon className="w-4 h-4" />
-              追加
-            </button>
-          )}
-        </div>
-
-        <div className="mt-4 space-y-4">
-          <div>
-            <span className="block text-sm font-medium text-stone-700 mb-2">
-              配信タイミング
-            </span>
-            <div className="flex rounded-lg border border-stone-300 overflow-hidden">
-              <button
-                type="button"
-                onClick={() => setTiming("now")}
-                className={`flex-1 px-4 py-2.5 text-sm font-medium transition-colors flex items-center justify-center gap-1.5 ${
-                  timing === "now"
-                    ? "bg-teal-600 text-white"
-                    : "bg-white text-stone-600 hover:bg-stone-50"
-                }`}
-              >
-                <PaperAirplaneIcon className="w-4 h-4" />
-                今すぐ送信
-              </button>
-              <button
-                type="button"
-                onClick={() => setTiming("schedule")}
-                className={`flex-1 px-4 py-2.5 text-sm font-medium transition-colors border-l border-stone-300 flex items-center justify-center gap-1.5 ${
-                  timing === "schedule"
-                    ? "bg-blue-600 text-white"
-                    : "bg-white text-stone-600 hover:bg-stone-50"
-                }`}
-              >
-                <ClockIcon className="w-4 h-4" />
-                スケジュール
-              </button>
-            </div>
-          </div>
-
-          {timing === "schedule" && (
-            <div>
-              <label
-                htmlFor="broadcast-scheduled"
-                className="block text-sm font-medium text-stone-700 mb-1"
-              >
-                配信日時
-              </label>
-              <input
-                id="broadcast-scheduled"
-                type="datetime-local"
-                value={scheduledAt}
-                onChange={(e) => setScheduledAt(e.target.value)}
-                className="w-full px-3 py-2 border border-stone-300 rounded-lg text-sm focus:outline-none focus:ring-2 focus:ring-teal-500 focus:border-transparent"
-              />
-            </div>
-          )}
-        </div>
-
-        {isError && (
-          <div className="mt-4 px-4 py-3 rounded-lg text-sm bg-red-50 text-red-700">
-            {errorMessage || "エラーが発生しました"}
-          </div>
-        )}
-
-        <div className="mt-6 flex justify-end">
-          <button
-            type="button"
-            onClick={handleSubmit}
-            disabled={!isValid || isPending}
-            className={`px-6 py-2.5 text-sm font-medium text-white rounded-lg disabled:opacity-50 disabled:cursor-not-allowed transition-colors flex items-center gap-1.5 ${
-              timing === "now"
-                ? "bg-teal-600 hover:bg-teal-700"
-                : "bg-blue-600 hover:bg-blue-700"
-            }`}
-          >
-            {isPending ? (
-              "送信中..."
-            ) : timing === "now" ? (
-              <>
-                <PaperAirplaneIcon className="w-4 h-4" />
-                送信する
-              </>
-            ) : (
-              <>
-                <ClockIcon className="w-4 h-4" />
-                予約する
-              </>
-            )}
-          </button>
-        </div>
-      </div>
-    </div>
-  );
-};
-
-const BroadcastPartPreview = ({
-  parts,
-}: {
-  parts: string | null;
-  body: string;
-}) => {
-  if (!parts) return null;
-
-  let parsed: BroadcastPart[];
-  try {
-    parsed = JSON.parse(parts);
-  } catch {
-    return null;
-  }
-
-  const hasImage = parsed.some((p) => p.type === "image");
-  if (!hasImage) return null;
-
-  return (
-    <div className="flex gap-1.5 mb-1.5">
-      {parsed
-        .filter(
-          (p): p is Extract<BroadcastPart, { type: "image" }> =>
-            p.type === "image",
-        )
-        .map((p) => (
-          <img
-            key={p.imageR2Key}
-            src={getImageUrl(p.imageR2Key)}
-            alt=""
-            className="w-12 h-12 rounded object-cover border border-stone-200"
-          />
-        ))}
-    </div>
-  );
 };
 
 export const BroadcastPanel = () => {
@@ -581,10 +195,7 @@ export const BroadcastPanel = () => {
                           : `作成: ${formatDateTime(broadcast.createdAt)}`}
                     </span>
                   </div>
-                  <BroadcastPartPreview
-                    parts={broadcast.parts}
-                    body={broadcast.body}
-                  />
+                  <BroadcastPartPreview parts={broadcast.parts} />
                   <p className="text-sm text-stone-700 line-clamp-3 whitespace-pre-wrap">
                     {broadcast.body}
                   </p>
